merge feature-be/orderMVC to develop

See merge request sw_track/class_05/web_project/team14/fragrant!52

+++ src/controllers/orderController.js
@@ -6,7 +6,7 @@ const orderController = {
     try {
       const { products, orderer, price, orderStatus, requirement } = req.body;
       const { userEmail } = req;
-      await orderService.createOrder(
+      const isCreated = await orderService.createOrder(
         {
           products,
           orderer,
@@ -16,7 +16,10 @@ const orderController = {
         },
         userEmail,
       );
-      await addProductsQuantity(products);
+      if (!isCreated) {
+        throw new Error('주문에 실패했습니다.');
+      }
+      await subtractProductsQuantity(products);
       res.status(201).end();
     } catch (err) {
       next(err);
@@ -27,6 +30,9 @@ const orderController = {
     try {
       const { orderId } = req.params;
       const order = await orderService.getOrder(orderId);
+      if (!order) {
+        throw new Error('상품 조회에 실패했습니다.');
+      }
       res.json(order);
     } catch (err) {
       next(err);
@@ -36,6 +42,9 @@ const orderController = {
   async getOrders(req, res, next) {
     try {
       const orders = await orderService.getOrders();
+      if (!orders) {
+        throw new Error('상품 리스트 조회에 실패했습니다.');
+      }
       res.json(orders);
     } catch (err) {
       next(err);
@@ -46,6 +55,9 @@ const orderController = {
     try {
       const { userName } = req.query;
       const orders = await orderService.getOrdersByUserName(userName);
+      if (!orders) {
+        throw new Error('상품 리스트 조회에 실패했습니다.');
+      }
       res.json(orders);
     } catch (err) {
       next(err);
@@ -56,6 +68,9 @@ const orderController = {
     try {
       const { userEmail } = req.user;
       const orders = await orderService.getOrdersByUserEmail(userEmail);
+      if (!orders) {
+        throw new Error('상품 리스트 조회에 실패했습니다.');
+      }
       res.json(orders);
     } catch (err) {
       next(err);
@@ -69,7 +84,10 @@ const orderController = {
       orderer.email = req.userEmail;
       const toUpdate = { orderer };
       if (requirement) toUpdate.requirement = requirement;
-      await orderService.updateUserOrder(orderId, toUpdate);
+      const isUpdated = await orderService.updateUserOrder(orderId, toUpdate);
+      if (!isUpdated) {
+        throw new Error('주문 내역 수정에 실패했습니다.');
+      }
       res.status(201).end();
     } catch (err) {
       next(err);
@@ -80,7 +98,10 @@ const orderController = {
     try {
       const { orderId } = req.params;
       const { orderStatus } = req.body;
-      await orderService.updateUserOrder(orderId, { orderStatus });
+      const isUpdated = await orderService.updateUserOrder(orderId, { orderStatus });
+      if (!isUpdated) {
+        throw new Error('주문 내역 수정에 실패했습니다.');
+      }
       res.status(201).end();
     } catch (err) {
       next(err);
@@ -90,9 +111,12 @@ const orderController = {
   async deleteOrder(req, res, next) {
     try {
       const { orderId } = req.params;
-      await orderService.deleteOrder(orderId);
       const order = await orderService.getOrder(orderId);
-      await subtractProductsQuantity(order.products);
+      if (!order) {
+        throw new Error('취소할 주문이 없습니다.');
+      }
+      await orderService.deleteOrder(orderId);
+      await addProductsQuantity(order.products);
       res.end();
     } catch (err) {
       next(err);
